[#47] refactor : #46 PR에 기반하여 리팩토링 및 합병
