+++ src/pages/User/RegisterInterestPage.tsx
@@ -1,13 +1,9 @@
 // 개인정보등록 : 관심사
+import styled from "styled-components";
 import { useState } from "react";
 import { SelectInterest, SelectInterestResult } from "components/User/RegisterInterest";
-import { USER } from "@/util/constants";
+import { INIT_INDEX } from "@/util/constants";
 import { TSelectedInfo } from "@/util/types";
-import * as S from './style';
-
-const {
-  RegisterInterest: { INIT_INDEX },
-} = USER;
 
 const RegisterInterestPage = () => {
   const [selectedInfo, setSelectedInfo] = useState<TSelectedInfo>({
@@ -17,20 +13,40 @@ const RegisterInterestPage = () => {
   });
 
   return (
-    <S.InterestLayout>
+    <InterestLayout>
       <div className="interest__inner">
-        <S.InterestRow>
+        <InterestRow>
           <SelectInterest {...{ selectedInfo, setSelectedInfo }} />
-        </S.InterestRow>
+        </InterestRow>
 
         {selectedInfo.items.length > 0 && (
-          <S.InterestRow>
+          <InterestRow>
             <SelectInterestResult {...{ selectedInfo, setSelectedInfo }} />
-          </S.InterestRow>
+          </InterestRow>
         )}
       </div>
-    </S.InterestLayout>
+    </InterestLayout>
   );
 };
 
 export default RegisterInterestPage;
+
+const InterestLayout = styled.div`
+  position: relative;
+  min-height: 100vh;
+  ${({ theme }) => theme.flexSet()}
+
+  .interest__inner {
+    height: calc(100vh / 2);
+
+    ${({ theme: { media } }) => media.desktop} {
+      font-size: 1.2rem;
+    }
+  }
+`;
+
+const InterestRow = styled.div`
+  position: relative;
+  height: fit-content;
+  padding: 3vh 0;
+`;
