+++ src/pages/index.tsx
@@ -1,15 +1,15 @@
 // pages : index
-import EnterPage from './EnterPage';
-import LoginPage from './auth/LoginPage';
-import RegisterInterestPage from './user/RegisterInterestPage';
-import RegisterLocationPage from './user/RegisterLocationPage';
-import HostRegistrationPage from './HostRegistrationPage';
-import MainPage from './MainPage';
-import MyInfoPage from './MyInfoPage';
-import StudyGroupingDetailPage from './StudyGroupingDetailPage';
-import RecruitmentsPage from './RecruitmentsPage';
-import RoomBookingPage from './RoomBookingPage';
-import RoomDetailPage from './RoomDetailPage';
+import EnterPage from "./EnterPage";
+import LoginPage from "./auth/LoginPage";
+import RegisterInterestPage from "./user/RegisterInterestPage";
+import RegisterLocationPage from "./user/RegisterLocationPage";
+import HostRegistrationPage from "./HostRegistrationPage";
+import MainPage from "./MainPage";
+import MyInfoPage from "./MyInfoPage";
+import StudyGroupingDetailPage from "./StudyGroupingDetailPage";
+import RecruitmentsPage from "./RecruitmentsPage";
+import RoomBookingPage from "./RoomBookingPage";
+import RoomDetailPage from "./RoomDetailPage";
 
 export {
   EnterPage,
