+++ src/util/constants.ts
@@ -10,6 +10,7 @@ export const ROUTE = {
   OAUTH_LOGOUT: "/oauth/kakao/logout",
 };
 
+// 이 부분도 컴포넌트에 따라 분리할까 생각중..
 export const TARGET_COUNT = 2;
 export const INIT_INDEX = -1;
 export const MAX_SELECT_NUM = 5;
