+++ src/util/types/index.ts
@@ -1,5 +1,5 @@
 import { TChildren, TFont, TPage } from "./normal";
-import { TSelectedInfo, TInterestSelectProps } from "./user";
+import { TSelectItem, TSelectedInfo, TInterestSelectProps, TInterestDetailProp, TInterestInfo } from "./user";
 
-export type { TChildren, TFont, TPage };    // normal
-export type { TSelectedInfo, TInterestSelectProps };    // user
+export type { TChildren, TFont, TPage }; // normal
+export type { TSelectItem, TSelectedInfo, TInterestSelectProps, TInterestDetailProp, TInterestInfo }; // user
